requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
arabic-reshaper==3.0.0
asgiref==3.8.1
brazilcep==7.0.0
Brotli==1.1.0
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
cryptography==45.0.2
cssbeautifier==1.15.4
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-localflavor==4.0
django-phonenumber-field==8.1.0
django-widget-tweaks==1.5.0
djlint==1.36.4
EditorConfig==0.17.0
fonttools==4.58.0
frozenlist==1.6.0
html5lib==1.1
jsbeautifier==1.15.4
json5==0.12.0
lxml==5.4.0
multidict==6.4.3
oscrypto==1.3.0
pathspec==0.12.1
phonenumbers==9.0.5
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-stdnum==2.0
PyYAML==6.0.2
qrcode==8.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
yarl==1.20.0
zopfli==0.2.3.post1
